feat: implement local data source for tasks, and restructure for categories
