+++ app/src/main/java/com/washingtondcsquad/tudee/data/services/CategoriesServiceImpl.kt
@@ -1,5 +1,6 @@
 package com.washingtondcsquad.tudee.data.services
 
+import com.washingtondcsquad.tudee.data.dataSources.CategoryLocalDataSource
 import com.washingtondcsquad.tudee.data.localSource.mapper.category.toDomain
 import com.washingtondcsquad.tudee.data.localSource.mapper.category.toEntity
 import com.washingtondcsquad.tudee.domain.entity.Category
