+++ app/src/main/java/com/washingtondcsquad/tudee/data/utils/DataBaseConstants.kt
@@ -2,4 +2,5 @@ package com.washingtondcsquad.tudee.data.utils
 
 object DataBaseConstants {
     const val ROOM_DATABASE_NAME = "TudeeDatabase"
+    const val TASK_TABLE_NAME = "task_table"
 }
